Fixed bug with spaces in track names. Added map length line to analysis graph, ensured the finish is always showing. RNG updated to ull. Correctly show graph if the map is the same as what is from the save.

+++ NEA/RNG.h
@@ -8,7 +8,7 @@ namespace RNG
 	class Random : public std::mt19937
 	{
 	private:
-		unsigned int m_calls = 0;
+		unsigned long long m_calls = 0;
 		unsigned int m_seed;
 
 	public:
@@ -19,7 +19,7 @@ namespace RNG
 			return static_cast<std::mt19937*>(this)->operator()();
 		}
 
-		const unsigned int GetCalls()
+		const unsigned long long GetCalls()
 		{
 			return m_calls;
 		}
